Revert "Some improvments"
This reverts commit fe7b4f953fba8aadf6ff679c59c6d07d978d281b.

+++ commands/voice.js
@@ -163,9 +163,7 @@ module.exports = {
     log.info(member2name(before.member, 1, 1), 'delete', '#' + before.channel.name);
     try{
         before.channel.delete();
-    } catch {
-        console.log('Unknown channel')
-    };
+    } catch {console.log('Unknown channel')}
   },
 
 
@@ -258,13 +256,7 @@ module.exports = {
       const permission = this.channel.permissionOverwrites.get(user[0].id);
       if(permission) permission.delete();
       const text = channel.guild.channels.cache.get(user[0].text_id);
-      if(text) {
-        try{
-            text.delete();
-        } catch {
-            console.log('Unknown channel: ' + text)
-        };
-        };
+      if(text) text.delete();
     }
 
     const voice = {
